move types to catalog-common
