+++ plugins/catalog-common/src/ingestion/RecursivePartial.ts
@@ -14,8 +14,14 @@
  * limitations under the License.
  */
 
-export type {
-  LocationAnalyzer,
-  ScmLocationAnalyzer,
-  AnalyzeOptions,
-} from './types';
+/**
+ * Makes all keys of an entire hierarchy optional.
+ * @ignore
+ */
+export type RecursivePartial<T> = {
+  [P in keyof T]?: T[P] extends (infer U)[]
+    ? RecursivePartial<U>[]
+    : T[P] extends object
+    ? RecursivePartial<T[P]>
+    : T[P];
+};
