+++ plugins/catalog-common/src/ingestion/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2022 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,7 @@
  */
 
 export type {
-  LocationAnalyzer,
-  ScmLocationAnalyzer,
-  AnalyzeOptions,
-} from './types';
+  AnalyzeLocationResponse,
+  AnalyzeLocationRequest,
+  AnalyzeLocationExistingEntity,
+} from './LocationAnalyzer';
